+++ .devcontainer/on-create.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+# run dotnet restore
+dotnet restore src/webvalidate.sln
+
+# copy vscode files
+mkdir -p .vscode && cp .vscode-template/* .vscode