+++ src/app/WebValidation/Model/config.cs
@@ -36,7 +36,7 @@ namespace CSE.WebValidate
         /// <summary>
         /// XML
         /// </summary>
-        Xml,
+        Xml
     }
 
     /// <summary>